Subiendo proyecto a github

+++ backend/routes/products.js
@@ -7,28 +7,27 @@ const router = require("express").Router();
 //CREATE
 
 router.post("/", isAdmin, async (req, res) => {
-  const { name, brand, desc, price, image } = req.body;
+  const { name, category, details, price, image } = req.body;
 
   try {
-    if (image) {
-      const uploadedResponse = await cloudinary.uploader.upload(image, {
-        upload_preset: "online-shop",
-      });
+    const uploadedResponse = await cloudinary.uploader.upload(image, {
+      upload_preset: "online-shop",
+    });
 
-      if (uploadedResponse) {
-        const product = new Product({
-          name,
-          brand,
-          desc,
-          price,
-          image: uploadedResponse,
-        });
 
-        const savedProduct = await product.save();
-        res.status(200).send(savedProduct);
-      }
-    }
-  } catch (error) {
+    const product = new Product({
+      name,
+      category,
+      details,
+      price,
+      image: uploadedResponse,
+    })
+
+    const savedProduct = await product.save();
+    res.status(200).send(savedProduct);
+  }
+
+  catch (error) {
     console.log(error);
     res.status(500).send(error);
   }
@@ -36,10 +35,24 @@ router.post("/", isAdmin, async (req, res) => {
 
 //DELETE
 
-router.delete("/:id", isAdmin, async (req, res) => {
+router.delete("/:id", async (req, res) => {
   try {
-    await Product.findByIdAndDelete(req.params.id);
-    res.status(200).send("Product has been deleted...");
+    const product = await Product.findById(req.params.id);
+
+    if (!product) return res.status(404).send("Product not found... ");
+
+    if (product.image.public_id) {
+      const destroyResponse = await cloudinary.uploader.destroy(
+        product.image.public_id
+      );
+      if (destroyResponse) {
+        const deletedProduct = await Product.findByIdAndDelete(req.params.id);
+
+        res.status(200).send(deletedProduct);
+      }
+    } else {
+      console.log("Action terminated. Failed to deleted product image ... ");
+    }
   } catch (error) {
     res.status(500).send(error);
   }
@@ -48,13 +61,13 @@ router.delete("/:id", isAdmin, async (req, res) => {
 //GET ALL PRODUCTS
 
 router.get("/", async (req, res) => {
-  const qbrand = req.query.brand;
+  const name = req.query.brand;
   try {
     let products;
 
-    if (qbrand) {
+    if (name) {
       products = await Product.find({
-        brand: qbrand,
+        name: name,
       });
     } else {
       products = await Product.find();
@@ -81,14 +94,43 @@ router.get("/find/:id", async (req, res) => {
 
 router.put("/:id", isAdmin, async (req, res) => {
   try {
-    const updatedProduct = await Product.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: req.body,
-      },
-      { new: true }
-    );
-    res.status(200).send(updatedProduct);
+    if (req.body.productImg) {
+      const destroyResponse = await cloudinary.uploader.destroy(
+        req.body.product.image.public_id
+      )
+
+      if (destroyResponse) {
+        const uploadedResponse = await cloudinary.uploader.upload(req.body.productImg, {
+          upload_preset: "online-shop",
+        });
+
+        if (uploadedResponse) {
+          const updatedProduct = await Product.findByIdAndUpdate(
+            req.params.id,
+            {
+              $set: {
+                ...req.body.product,
+                image: uploadedResponse,
+              }
+            }, { new: true }
+          )
+          res.status(200).send(updatedProduct);
+        }
+      }
+
+    }
+
+    else {
+      const updatedProduct = await Product.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            ...req.body.product,
+          }
+        }, { new: true }
+      )
+      res.status(200).send(updatedProduct);
+    }
   } catch (error) {
     res.status(500).send(error);
   }
